(Fix):moved context.jsx and done changes in index.jsx
