+++ src/components/content.jsx
@@ -1,6 +1,7 @@
 import CardData from '../config/CardData.mjs';
 import CategoryCards from './cards/CategoryCards.jsx';
-import underline from '../../assets/underline .svg';
+import underline from '../assets/underline .svg'
+
 const Content = () => {
   const card = CardData.map((value) => {
     return <CategoryCards key={value.id} card={value} />;
