+++ src/pages/LandingPage/index.jsx
@@ -1,7 +1,8 @@
 import NavBar from '../../components/NavBar';
 import HeroSection from '../../components/HeroSection';
-import Content from './content';
+import Content from '../../components/content';
 import Footer from '../../components/Footer';
+
 
 const LandingPage = () => {
   return (
